Update is_null_terminated tests
The `is_null_terminated` tests that test the from_iso_array strings have
been updated to reflect the apparent reality of preallocation that takes
place when the length of the length of the array literal is not a power
of two. It appears as though the array does get pre-allocated with zero
bytes up to the next power of two, and as such, is null-terminated.