hashmap: Add new putindex function and make everything use it (#1432)
This change was motivated by profiling using linux `perf` and
`FlameGraphs` which revealed that `ponyint_objectmap_getorput`
was an expensive operation of which `search` in hashmap was taking
up almost all the processing time. The `FlameGraphs` also revealed
that `search` was being called twice by `ponyint_objectmap_getorput`,
once for `get` and again for `put` where the object wasn't found.
This change allows `ponyint_objectmap_getorput` to only do the `search`
once and reuse the `index` returned by `search` for the subsequent `put`
operation via the new `putindex` function. In addition to changing
`ponyint_objectmap_getorput`, this PR also changes all instances
where a `get` was followed by a subsequent `put` to reuse the `index`
from `search` via `putindex` for all hashmaps in the runtime and the
compiler.