Prevent `String.truncate` from chopping last byte
When `String.truncate` is called on a String where `_size` is
to `_alloc`, the last byte would be lost since it set the null
terminator without first checking.

Because the docs guarantee that the result is null terminated,
when the `len` passed to `truncate` is greater than or equal to
the `_alloc`, extra space is reserved to accommodate the null
terminator.

While this resolves the issue without changing documented behavior,
it has the undesirable effect that in this specific case, a call to
`truncate` can actually allocate, which may be surprising.