Add ASIO one shot support when using epoll
Using asio one shot can be much more efficient than the current edge
polling. We started using at Sendence with TCPConnection and saw a
large drop in the number of events being sent to TCPConnection
instances. This in turn lead to performance improvements.

This commit contains support in the linux asio code, as well as
the linux tcp socket code to use one shot.

Support for OSX/FreeBSD can be added as kqueue supports the same
fundamental options.