Improve efficiency of muted TCPConnection on non Windows platforms (#1477)
Prior to this commit, if we were in TCPConnection's _pending_reads
and encountered discovered that the connection was muted, we would
call read_again to requeue another read and would then return.

This results in a lot of extra, unneeded work. Until the connection
is unmuted, there's no point to calling read_again, we'll just keep
repeating the same cycle over and over again; pointlessly putting
messages on the TCPConnection's queue.

With the commit, we now return from _pending_reads when discovering
that the connection is muted without calling read_again. Instead,
when we unmute the connection, we will then immediately call
_pending_reads to start processing any unhandled data.

There is probably a similar improvement that can be made on Windows,
however, I'm not a Windows programmer, barely know IOCP at all and
am more likely to break the Windows TCP implementation than I am
improve it therefore, this commit only applies to non Windows platforms.