Linux epoll event resubscribe optimization
Prior to this change, if ONE_SHOT is enabled, linux epoll based
asio wouldn't be subscribed for certain events after another
event was fired on the same FD.

The specific scenario was the following:

* Register event E for READ/WRITE with ONE_SHOT.
* E fires for READ and a message is placed on the message queue
  for the TCPConnection actor.
* TCPConnection actor eventually gets to processing this message
  (after processing any other messages ahread of this one on the
  queue) and then goes through the work of `_event_notify`. At
  the end of `_event_notify`, `_resubscribe_event` is called and
  epoll is told to resubscribe for READ and/or WRITE events as
  required.

Normally, this wouldn't be an issue but the behavior we're seen
is that with ONE_SHOT, if a READ (or WRITE) event is fired, all
events for the FD are disabled. This means that in the above
scenario, when the READ event is fired, the WRITE is also
disabled along with the READ because of ONE_SHOT.

We could (and did) try adding a call to `_resubscribe_event` in
`_event_notify` before any processing is done to resubscribe
earlier but there were two problems with this. First, it still
meant that the other event was unsubscribed for the duration of
the time the asio event notification message was sitting on the
message queue of the TCPConnection actor. Second, it resulted in
a race condition and incorrect use of epoll where the TCPConnection
actor would ask epoll to resubscribe to either READ or WRITE when
the event had already fired and was waiting on the message queue.
For the READ event case, this is incorrect use of epoll because
the manual for epoll strongly indicates that when using ONE_SHOT
and Edge Triggering all data should be read from the socket before
it is resubcribed again.

The final design we settled upon is to move the readable/writeable
flags into the Event structure, split the resubscribe for READ
and WRITE events, and have epoll automagically resubscribe to the
untriggered event if ONE_SHOT is enabled. This ensures there are
no race conditions because the decision to resubscribe is based on
the the state of the readable/writeable flags on the event which is
always up to date based on the latest events that fired. It also
ensures that we resubscribe to the untriggered event as soon as
possible because the epoll asio thread takes care of it.

* The event now holds the readable/writeable flags.
* This is only safe because the asio thread will only
  write to those flags if there's a subscription for an event.
* In order to avoid two threads writing/reading to the new flags
  at the same time, the resubscribe for reads and writes has been
  split.
* Additionally, the epoll code will now automagically resubscribe
  to the un-triggered event in order to minimize the delay for when
  an event is triggered for ONE_SHOT code so it behaves the same
  as for kqueue ONE_SHOT where if an FD is subscribed with ONE_SHOT
  for both `read` and `write` events and a `read` event fires and
  is disabled because of ONE_SHOT, the `write` event is still left
  enabled. This also greatly simplifies where/when resubscribes
  need to occur in the TCPConnection actor.
* Update kqueue code to also split readable/writeable resubscribe.