Slightly more efficient work stealing (#1505)
In the case where choose_victim returns NULL, we set the
scheduler trying to do stealing as the victim. This caused
it to check the inject queue and then its own queue for work.
The latter check is pointless. We are doing work stealing
because our queue is empty.

This commit switches to only checking the inject queue if
we are "stealing from ourselves".