Prevent type parameter shadowing of builtin types
When a type parameter is introduced which matches any of the simple UIF
names, it causes any other such simple type to be recognized as a type
parameter reference, which results in the compiler aborting.

In the discussion on ponylang/ponyc#1526, it was decided that all
shadowing by type parameters should be prohibited.

Hence, in this patch, if we're resolving a type param whose id is still
nominal, we check the nearest module to see if that name is already
defined.