Correctly detect compile errors from arithmetic operations
Previously in `special_case_operator`, a no-op and a failure were
reported in the same way. This change ensures that we detect errors and
that we don't proceed through the final stages of compilation.

This doesn't fix any program correctness issue. The only error that
could be detected here is a division by 0 error which was reported even
though compilation didn't stop and wasn't inhibiting the runtime check,
because of the way the code generator is architectured.