Trace message contents with the refcap of the receiver
In a previous change (f9591b2), the tracing algorithm was changed to
trace objects in the context of the sender. Although it is an
optimisation regarding object types, it is a pessimisation regarding
refcaps. This change resolves the issue by tracing with the sender's
type context and the receiver's refcap context.