Load program symbols dynamically in the runtime
The __DescTable and __DescTableSize defined by Pony programs and
generated by the compiler are now accessed dynamically by the runtime
(with dlsym on POSIX and GetProcAddress on Windows).

This change was made for two reasons:

- To avoid link errors from of compiler-generated symbols when using
  the runtime library directly in a C program.
- JIT compilation is being worked on for the compiler test framework.
  Since we'll want to run more than one JIT'ed programs in the test
  suite, the symbol addresses must be dynamic.

As part of this change, the pony_start runtime function now takes an
additional parameter indicating whether the language-specific features
of the Pony runtime should be intialised. This solves the first concern
stated above by avoiding referencing the symbols.