Performance improvement in subtype checking.
In subtype checking we now avoid some costly allocation operations
associated with viewpoint adaptation and reification by checking
first if the type could be a subtype of the other if caps were ignored.
If not, then there's no point in moving forward with comparisons
that take the cap into account, but require costly allocations.

Reduces compile time drastically for programs requiring a lot of type
unions to be flattened, where those types involve viewpoint adaptation.
For the real-world package under test, compile time was reduced by 23x.
For comparison, standard library test compile time was not affected by this change.