Makes NetAddress comparable
Provides two types of equality.

1: "basic equality"

Compares all the fields of NetAddress to see if they are equal. This
is the comparison if we use `==` or `eq`.

2: "address equality"

Compares based on the ip address portion of NetAddress. Does not
take port, family, length or scope into account. We are using
"address equality" at Sendence to compare, "is this IP for a local
or remote address" which is in turn used to decide whether to turn
on TCP_NODELAY (on Linux, using TCP_NODELAY on local addresses has
a negative impact on our performance/stability. Using it on
remote addresses helps our performance/stability).