Hashmap optimization for resize
Change hashmap to not have to zero out all memory on init
and resize but only the item bitmap.

This required changing all element validity tests to use
the item bitmap instead of checking if `ptr` is NULL or not.

Change integer division/module for item_bitmap access to use
bitwise operations instead based on @sylvanc's suggestion.

Change `search` to avoid some memory accesses when a bucket
is empty based on @sylvanc's suggestion.