Change `cpointer` in `array` and `string` to take offset
Prior to this change getting an offset for a pointer into an
`array` or a `string` required doing pointer arithmetic after
geting the pointer to the start. This is problematic because
in most cases we want the pointer to a specific entry in the
`array` or `string`. Doing pointer arithmetic to calculate the
pointer to a specific entry is non-trivial because if requires
one to know the size of each array entry in order to do the
pointer math correctly.

This change makes it easy and safe to get the pointer to a
specific entry by giving an optional `offset` argument to
`cpointer` in `array` and `string` and letting Pony handle
the appropriate pointer math prior to returning the pointer
to the appropriate element.