Add custom assertion mechanism with backtrace printing
This change will allow getting more complete reports from assertion
failures. This will smooth the the bug report process, as users won't
have to get a backtrace from a debugger. As part of this change,
assertions are now enabled in release builds of the compiler, as
discussed in #1553. They still are disabled in release builds of the
runtime.

In the future, this system could be extended with even more precise
information like line numbers. Parsing the debug information of the
executables and libraries will be required for this.

Closes #1553.