Fix message sending of boxed types through interfaces (#1596)
Previously, a message send was inlined in the caller when every
possible instantiation of the method was a behaviour. This logic turned
out to be incorrect when sending a message through an interface, with
the interface expecting an unboxed value as a parameter and one of the
subtypes expecting a boxed value. This change suppresses the inlining
of message sends when such a case arises.