Pick up changes to the exit code from primitive finalisers (#1613)
Previously, a primitive finaliser calling `pony_exitcode` wouldn't
affect the actual exit code of the program. This occurred because the
value used was the value returned from `pony_start`, and finalisers
were ran after returning from `pony_start`.

Also, the language-related parts of the runtime such as sockets are now
correctly finalised when not running the runtime as a library.

This change adds the `pony_get_exitcode` function, which returns the
current value of the exit code as set by `pony_exitcode`.