Change File class to use low level and vectored i/o functions (#1598)
* Change File class to use low level i/o functions

The motivation behind this is to improve performance to file i/o
operations. The C standard stdio functions do user space
buffering prior to sending the data to the OS kernel via a system
call. On most operating system, the OS kernel also does additional
buffering on top of the user level buffering by the C functions.
Unfortunately, if the C functions are buffering data, they prevent
the OS kernal from taking advantage of some potential performance
enhancements if they're available (namely Vectored I/O so the
OS can offload work to hardward and never has to combine multiple
user buffers into a single buffer for writing to disk).

The commit changes the File class to use low level i/o functions
that map to system calls on Linux and other posix based systems
and to C lib calls on Windows. On Posix based systems, the OS is
free to take advantage of performance offloading to hardware if
possible. On Windows, this change shouldn't have a negative
performance impact but doesn't give the same possible performance
benefits as on Posix based systems.

Other changes:

* Enable `_FILE_OFFSET_BITS=64` arg to compiler to ensure 32 bit
  linux builds get large file support enabled.
* Use `lseek64` and `ftruncate64` on linux to ensure 32 bit builds
  are using 64 bit versions of the functions for large file
  support.
* Adds `datasync` to be able to call `fdatasync` instead of `fsync`
  as it can be appropriate in some situations. On Windows, `fsync`
  and `fdatasync` behave the same.
* Adds `queue` and `queuev` functions that are analogous to
  `write` and `writev` but queue/buffer data to be written
  instead of writing the data immediately. Additionally,
  these new functions provide the ability to interleave buffered
  data writing and unbuffered data writing where unbuffered data
  writing would force any buffered data to be written first.
* Changes `flush` to flush any queued/buffered data to disk
  because it is otherwise unnecessary since we're not buffering
  data in the C library via stdio any longer.
* A few additional tests have been added to validate things are
  working correctly.
* Cast the relevant `rlim_t` to `size_t` as required to ensure
  things compile correctly on 32 bit platforms. There should be
  no negative consequences to losing the extra bits becuase stack
  sizes (for which we use `rlimit`) should never be larger than a
  32 bit size on a 32 bit platform. The cast to `size_t` has no
  negative impact on 64 bit platforms where `size_t` and `rlim_t`
  are both 64 bits already.

Future work:
* Change `queue`/`queuev` to automatically flush data to disk once
  the # of queued items exceeds `IOV_MAX` for a single system call.
* Error handling in the File package might be possible to add now
  including having the various functions throw `error` on things
  that can't be recovered from.
* Possibly switch to async/nonblocking i/o for file writing with an
  actor as the intermediary (similar to the TCPConnection actor in
  the net package).

* Change `cpointer` in `array` and `string` to take offset

Prior to this change getting an offset for a pointer into an
`array` or a `string` required doing pointer arithmetic after
geting the pointer to the start. This is problematic because
in most cases we want the pointer to a specific entry in the
`array` or `string`. Doing pointer arithmetic to calculate the
pointer to a specific entry is non-trivial because if requires
one to know the size of each array entry in order to do the
pointer math correctly.

This change makes it easy and safe to get the pointer to a
specific entry by giving an optional `offset` argument to
`cpointer` in `array` and `string` and letting Pony handle
the appropriate pointer math prior to returning the pointer
to the appropriate element.