Change benchmarks to limit use of `PauseTiming`/`ResumeTiming`
According to the Google Benchmark library issue:

https://github.com/google/benchmark/issues/179

The `PauseTiming`/`ResumeTiming` functions introduce a large
amount of overhead and are not recommended to be used for
microbenchmarks without understanding the overheads involved.

This commit is an attempt to minimise use of these functions along
with adding a disclaimer for when these functions are used and
a link to the issue so everyone can understand the implications.

It also changes names of some benchmarks in order to make it easy
to visually identify benchmarks that use these functions.