Use unsafe operations in persistent map (#1618)
This PR replaces some math operations in the `_MapNode` methods
`_set_bit`, `_clear_bit`, `_has_bit`, and `_mask` with the unsafe
versions. This change should not cause issues because the argument
`i` is never greater than 31 and the argument `l` is never greater
than 6. The purpose of this change is to improve the overall
performance of the persistent map.