Push local pool memory to a global list at thread termination
This change adds a global list of free blocks to the pool allocator,
similar to the existing global list of free lists. When a runtime
thread terminates, it will push its local free blocks to the global
list, which can then be retreived by other threads.

The goal of this change is to avoid memory leaks when starting and
stopping the runtime multiple times in a program, for example in the
compiler test suite.

This pool allocator cleanup functionality is exposed to user threads
through the new pony_unregister_thread function.