Only update online API docs when releasing
This will keep the online API docs in sync with the last release.
This feels like a less surprising thing to happen then to
have the docs change with each commit to master.

We could invest more time in the future to keep multiple versions
of the docs online + latest build docs but that is more than I
have time for at the moment.

If someone wanted to take that on, I'd be happy to guide them.