Remove the check for union types on match error (#1630)
* Remove the check for union types on match error

A nice attempt to produce a much friendlier error message when a union
type pattern is involved and a match expression doesn't type check is
sadly removed in this commit.

Unfortunately, it assumed union types would be composed of nominal types
(not true), and that the problem with the match was a union type with
varying capabilities (this is actually allowed).

A nicer error message would indeed be useful here. The right approach
would be to emit errors in `matchtype.c` when they are needed, i.e. on
DENY.

* remove the associated test