Represent persistent map pairs as tuple (#1632)
* Represent persistent map pairs as tuple

Previously the persistent map stored each key-value pair as a class
along with a hash of the key. The pairs are now stored as a 2-tuple of
the key and value and the hash is no longer stored. The hash is only
needed on the creation of a new sub-node and is created in that
situation. Also, number type conversions from U32 to USize and from U8
to U32 are using the unsafe operations. These changes are for improved
performance of the persistent map.

* formatting

* add benchmarks for persistent map

* remove binary file