Refactor ast internal structure to store annotation more efficiently.
Now annotation and type are both stored, while only using one pointer
(and one flag) rather than two pointers on the ast_t struct.

This saves size in allocations, and particularly puts us in a smaller
size class when it comes to the next-power-of-two size that dictates
which memory pool we allocate from.