Close over free variables in lambdas and object literals (#1648)
Previously, using a variable from the surrounding lexical scope in an object literal or a lambda produced a compiler error. This was surprising to the programmer.

This change allows both lambdas and object literals to use variables in the surrounding lexical scope, as the programmer would expect. In both cases, the closure is expressed under the hood as adding a field that is initialised by passing an alias of the variable to the constructor. For example:

let x = "hi"
let f = {(): String => x}
Here, when f rewrites to an object literal, it becomes:

let f =
  object
    var x: String = x
    fun apply(): String => x
  end