Fix AVX/AVX2 issues with prebuilt ponyc
We regularly get issues opened from people having problems
with the prebuilt pony packages where their platform
doesn't support AVX or AVX2 and the prebuilt packages were
built with that on.

This commit should get us building with the simplest of baselines
for users and allow them to run without issue.

The `arch=` idea is Sylvan's.