Build Linux release binaries correctly
Before this change, when releasing to bintray, we would run

`make deploy arch=x86-64 ... deploy`

with the intent of building a pony compiler and runtime that targets
the x68-64 instruction set. Unfortunately that isn't what would happen.
Because we had previously run

`make test-ci`

Which doesnt specify an architecture and therefore uses features
available on the build machine, we were in fact uploading an arch
specific build- the one created by `make test-ci`.

With this change, we will remove all `make test-ci` related items and
then rebuild for release with the correct make settings.

Fixes #1696