Fix compiler assert on arrow to typeparam in constraint.
Fixes by moving "in constraint" checks for arrow and tupletype
from the flatten pass to the syntax pass, so that they are guaranteed
to happen before we try to set the cap on the typeparamref.

It's also more appropriate to run them there since they are syntax-level
checks that require no knowledge of name resolution, or anything else
beyond simple syntax frame lookups.

Resolves #1694.