Refactor control tracking to use AST_FLAG_JUMPS_AWAY instead of control types.
This lets us track that a type will "jump away" without conflating that with
the process of type checking, so we can decouple reference status tracking
from type checking (and move it out of the expr pass in a future commit).

This is a little bit conceptually cleaner, since we have stopped using
control flow tokens as types that represent "this expression has no type
because the control flow jumps away".

It also removes some allocations, since we're no longer allocating
type nodes for attaching to control types that jump away.