Refactor to flatten union types in the parser instead of flatten pass.
Note that this means we are now only flattening structurally, instead
of also folding types that are subtypes of eachother into eachother.

However, checking subtype relations is actually what we're trying to
avoid, since it can be quite expensive for large unions.

A quick check of building a union-heavy project I'm working on shows
that this change reduces compile times of that project from ~2m to ~30s.

This is one of the compiler performance improvements discussed in #1562.