Add "getting started with GitHub" info to CONTRIBUTING (#1720)
From time-to-time we have folks wanting to contribute who have never
contributed via GitHub before. The process can be very daunting. This
commit adds a basic overview of the process to our CONTRIBUTING doc.
Additionally, it links to more detailed instructions on GitHub's site as
well as giving information on how the potential contributor can reach
out to the Pony community for help contributing.

[skip ci]