Use per-size lists when cleaning up the pool allocator
The previous cleanup strategy (introduced in 43a6346) consisted in
pushing every free blocks to a global free list, including the per-size
free block. There was a problem with this strategy: blocks were pulled
from that global list in `pool_alloc_pages`, and that function was
always called with a requested allocation size greater than POOL_ALIGN.
This means that the small blocks would never be pulled from the global
list to be reused, resulting in lots of memory fragmentation and
increased memory usage throughout multiple runs of the runtime.

This commit fixes this problem by using per-size global free block
lists in addition to the global free block list for oversized blocks.
When a thread exhausts its free block, it will now try to pull from the
appropriate per-size global list before trying to allocate a new block
of size POOL_ALIGN.