Fix false positive test failure on 32 bits (#1749)
The DigestofObject test was failing because of a problem in the
`ptr_to_int` helper function. This function was casting a pointer to
intptr_t, resulting in an incorrect sign extension. The function now
casts to uintptr_t, with the correct behaviour of a zero extension.

This doesn't fix any actual correctness issue in Pony programs.

Closes #1747.