Arrays as sequences (#1741)
* Arrays as sequences

Previously, a comma ',' was used to separate array elements. In Pony, a
comma is the tuple operator, and as a result, an array appeared to be
composed of a tuple of elements.

This change defines an array as a sequence of elements instead. As a
result, no separator is needed when elements are on new lines, or a
semicolon ';' (i.e. the sequence operator) can be used when elements are
on the same line.

This change means that all instances of ',' in Pony indicate a tuple
(including, logically speaking, parameter lists, arguments, type
arguments, etc.) and all instances of ';' indicate a sequence.

* remove space

* update pony.g