Support XCode 8.3 and LLVM 3.9
Previously, we used llvm-ar-3.8 if it was available. This doesn't work
with XCode 8.3, which will issue a link error. Instead, we now pick up
the llvm-ar from the `llvm-config --bindir` directory if it is
available, and otherwise fall back.

As part of this change, the homebrew location for llvm@3.9's llvm-config
is tried if no LLVM_CONFIG is specified.