Don't ignore buffer length when printing
As a left-over from when Pony strings where null terminated,
pony_os_std_print wasn't checking the length of the passed in
buffer and would print the entire buffer and beyond until it
hit a null terminator.

This PR fixes that issue so that we will now correctly only
print `len` elements from the buffer.