Fix Linux release scripts
the make deploy step ends up removing all created ponyc binaries,
this means if make docs comes after that, it will fail.

this commit makes each of the steps explicitly handle dependencies.

in the case of "deploy" this means running `make clean` first to
verify a clean working environment before building for deploy.

in the case of "make docs" this means building a ponyc installation
to use for creating docs from.