Do not allow capability subtyping when checking constraint subtyping.
The compiler currently allows contravariant constraint for type
arguments. When doing so it allows capability subtyping.

In other words, it allows the following, where `X` is a type variable,
`λ` a capability and `DS` is a type name.
```
                λ' ≤ λ
----------------------------------------
fun m[X: DS λ]() ≤ fun m[X: DS λ']()
```

This is however unsound, as shown in the example below. Bar's `alias`
method is able to create an `X^` from a `X!`, since `X` is a tag.
However, Foo's `alias` method is callable with an iso argument.

As a result, this can be used to duplicate an iso reference, as shown in
`Main`
```pony
trait Foo
  fun alias[X: Any iso](x: X!) : X^
class Bar is Foo
  fun alias[X: Any tag](x: X!) : X^ => x
class Baz

actor Main
  new create(env: Env) =>
    let foo : Foo = Bar

    let x : Baz iso = Baz
    let x' : Baz iso = foo.alias[Baz iso](x)
```

Instead, when checking constraint subtyping, the compiler should not
allow capabilities to be subtyped. This is similar to how type arguments
are checked to be within the constraint.